various fixes & component splitting
